// frontend/web/src/services/API/proserve-wb-provisioning-api/models/AvailableProduct.ts
export interface AvailableProduct {
    /**
     * Unique ID of the product.
     * @type {string}
     * @memberof AvailableProduct
     */
    productId: string;
    /**
     * Name of the product.
     * @type {string}
     * @memberof AvailableProduct
     */
    productName: string;
    /**
     * Type of the product (WORKBENCH|VIRTUAL_TARGET).
     * @type {string}
     * @memberof AvailableProduct
     */
    productType: string;
    /**
     * Description of the product.
     * @type {string}
     * @memberof AvailableProduct
     */
    productDescription?: string;
    /**
     * Total number of provisioning times that have been reported.
     * @type {number}
     * @memberof AvailableProduct
     */
    totalReportedTimes?: number;
    /**
     * Average provisioning time of a product.
     * @type {number}
     * @memberof AvailableProduct
     */
    averageProvisioningTime?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof AvailableProduct
     */
    availableStages: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof AvailableProduct
     */
    availableRegions: Array<string>;
    /**
     * Software Tools available in the product.
     * @type {Array<string>}
     * @memberof AvailableProduct
     */
    availableTools?: Array<string>;
    /**
     * Operating System versions available for the product.
     * @type {Array<string>}
     * @memberof AvailableProduct
     */
    availableOSVersions?: Array<string>;
}